bot/requirements.txt
python-telegram-bot==21.6
playwright==1.49.1
python-dotenv==1.0.1
httpx==0.27.2
